+++ src/handler/sequence/extract.rs
@@ -1,3 +1,6 @@
+//! Extract sequences from a collection of sequence files.
+//!
+//! Support extraction of sequences based on matching IDs or regular expressions.
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicUsize, Ordering};
 
@@ -14,28 +17,50 @@ use crate::writer::sequences::SeqWriter;
 
 impl OutputPrint for Extract<'_> {}
 
+/// Extraction options
 pub enum ExtractOpts {
+    /// Use regular expression to match sequence IDs
     Regex(String),
+    /// Match sequence IDs from a list of IDs
     Id(Vec<String>),
     None,
 }
 
+/// Extract sequences from a collection of sequence files.
+/// See each type for available options.
 pub struct Extract<'a> {
+    /// Input format of the sequence files
     input_fmt: &'a InputFmt,
+    /// Extraction options
     opts: &'a ExtractOpts,
+    /// Data type of the sequences
     datatype: &'a DataType,
+    /// Output directory
+    output_dir: &'a Path,
+    /// Output format of the sequence files
+    output_fmt: &'a OutputFmt,
 }
 
 impl<'a> Extract<'a> {
-    pub fn new(opts: &'a ExtractOpts, input_fmt: &'a InputFmt, datatype: &'a DataType) -> Self {
+    /// Create a new `Extract` instance
+    pub fn new(
+        input_fmt: &'a InputFmt,
+        datatype: &'a DataType,
+        opts: &'a ExtractOpts,
+        output_dir: &'a Path,
+        output_fmt: &'a OutputFmt,
+    ) -> Self {
         Self {
-            opts,
             input_fmt,
             datatype,
+            opts,
+            output_dir,
+            output_fmt,
         }
     }
 
-    pub fn extract_sequences(&self, files: &[PathBuf], output: &Path, output_fmt: &OutputFmt) {
+    /// Extract sequences with matching IDs or regular expressions
+    pub fn extract_sequences(&self, files: &[PathBuf]) {
         let file_counts = AtomicUsize::new(0);
         let spin = utils::set_spinner();
         spin.set_message("Extracting sequences with matching IDs...");
@@ -44,10 +69,11 @@ impl<'a> Extract<'a> {
             let matrix = self.get_matrix(seq);
             if !matrix.is_empty() {
                 let header = self.get_header(&matrix);
-                let outname = files::create_output_fname(output, file, output_fmt);
-                let mut writer = SeqWriter::new(&outname, &matrix, &header);
+                let output_name =
+                    files::create_output_fname(self.output_dir, file, self.output_fmt);
+                let mut writer = SeqWriter::new(&output_name, &matrix, &header);
                 writer
-                    .write_sequence(output_fmt)
+                    .write_sequence(self.output_fmt)
                     .expect("Failed writing the output files");
                 file_counts.fetch_add(1, Ordering::Relaxed);
             }
@@ -55,28 +81,28 @@ impl<'a> Extract<'a> {
         spin.finish_with_message("Finished extracting sequences!\n");
         let counts = file_counts.load(Ordering::Relaxed);
         assert!(counts > 0, "No matching IDs found!");
-        self.print_output_info(&counts, output, output_fmt);
+        self.print_output_info(&counts, self.output_dir, self.output_fmt);
     }
 
-    fn get_matrix(&self, seqmat: SeqMatrix) -> SeqMatrix {
-        let mut matrix: SeqMatrix = IndexMap::new();
+    fn get_matrix(&self, matrix: SeqMatrix) -> SeqMatrix {
+        let mut new_matrix: SeqMatrix = IndexMap::new();
         match self.opts {
-            ExtractOpts::Regex(re) => seqmat.iter().for_each(|(id, seq)| {
+            ExtractOpts::Regex(re) => matrix.iter().for_each(|(id, seq)| {
                 let matched_id = self.match_id(id, re);
                 if matched_id {
-                    matrix.insert(id.to_string(), seq.to_string());
+                    new_matrix.insert(id.to_string(), seq.to_string());
                 }
             }),
-            ExtractOpts::Id(ids) => seqmat.iter().for_each(|(id, seq)| {
+            ExtractOpts::Id(ids) => matrix.iter().for_each(|(id, seq)| {
                 ids.iter().for_each(|match_id| {
                     if match_id == id {
-                        matrix.insert(id.to_string(), seq.to_string());
+                        new_matrix.insert(id.to_string(), seq.to_string());
                     }
                 })
             }),
             ExtractOpts::None => panic!("Please, specify a matching parameter!"),
         };
-        matrix
+        new_matrix
     }
 
     fn match_id(&self, id: &str, re: &str) -> bool {
@@ -104,13 +130,22 @@ impl<'a> Extract<'a> {
 
 #[cfg(test)]
 mod tests {
+
     use super::*;
 
     #[test]
     fn test_match_id() {
         let id = "Bunomys_penitus";
         let re = "(?i)(Penitus)$";
-        let extract = Extract::new(&ExtractOpts::None, &InputFmt::Fasta, &DataType::Dna);
+        let output_dir = Path::new("tests/files");
+        let output_fmt = OutputFmt::Fasta;
+        let extract = Extract::new(
+            &InputFmt::Fasta,
+            &DataType::Dna,
+            &ExtractOpts::None,
+            &output_dir,
+            &output_fmt,
+        );
         assert!(extract.match_id(id, re));
     }
 
@@ -118,7 +153,15 @@ mod tests {
     fn test_get_matrix_re() {
         let re = ExtractOpts::Regex(String::from("(?i)(celebensis)"));
         let file = Path::new("tests/files/complete.nex");
-        let extract = Extract::new(&re, &InputFmt::Nexus, &DataType::Dna);
+        let output_dir = Path::new("tests/files");
+        let output_fmt = OutputFmt::Fasta;
+        let extract = Extract::new(
+            &InputFmt::Nexus,
+            &DataType::Dna,
+            &re,
+            output_dir,
+            &output_fmt,
+        );
         let (seq, _) = SeqParser::new(file, extract.datatype).parse(extract.input_fmt);
         let matrix = extract.get_matrix(seq);
         assert_eq!(2, matrix.len());
@@ -128,7 +171,15 @@ mod tests {
     fn test_get_matrix_id() {
         let re = ExtractOpts::Id(vec![String::from("Taeromys_calitrichus_NMVZ27408")]);
         let file = Path::new("tests/files/complete.nex");
-        let extract = Extract::new(&re, &InputFmt::Nexus, &DataType::Dna);
+        let output_dir = Path::new("tests/files");
+        let output_fmt = OutputFmt::Fasta;
+        let extract = Extract::new(
+            &InputFmt::Nexus,
+            &DataType::Dna,
+            &re,
+            output_dir,
+            &output_fmt,
+        );
         let (seq, _) = SeqParser::new(file, extract.datatype).parse(extract.input_fmt);
         let matrix = extract.get_matrix(seq);
         assert_eq!(1, matrix.len());