+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Remove unnecessary mutability for conversion modules.
 - Partition handle moved to align module.
+- Output directory and output format arguments for extraction now requires in the new instance of the extraction struct.
 
 ## v0.21.3